Check the basics
First of all, it checks the output of the contents by utilizing the list, and creates a file according to the data being saved, enters it, and saves it. And finally, when it is finalized, it is saved to the database.
Now all that's left is to save it to the database, save it as an object, and output it from the list.